fixed adding and deleting genres on edit page

+++ functions/delete_genre_from_movie.php
@@ -4,12 +4,11 @@ include 'connection.php';
 if(!empty($_POST['genre_select']) && !empty($_POST['movie_id'])){
   $genre_id = $mysqli->escape_string($_POST['genre_select']);
   $movie_id = $mysqli->escape_string($_POST['movie_id']);
-  $search = $_POST['search'];
-  $option = $_POST['option'];
-  $sorting_option = $_POST['sorting-option'];
 
-  $deletion_query = $mysqli->query("DELETE FROM is_genres WHERE genre_id=$genre_id AND movie_id=$movie_id");
+  $genre_query = $mysqli->query("SELECT * FROM is_genres WHERE movie_id=$movie_id");
 
-  header("location: ../pages/edit_page.php?movie_id=$movie_id&search=$search&option=$option&sorting-option=$sorting_option");
+  if($genre_query->num_rows != 1){
+    $deletion_query = $mysqli->query("DELETE FROM is_genres WHERE genre_id=$genre_id AND movie_id=$movie_id");
+  }
 }
 ?>
